requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0
anthropic>=0.39.0
requests>=2.31.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
joblib>=1.3.0
transformers>=4.48.0
torch>=2.5.0
yfinance>=0.2.54
openbb>=4.4.0
